Revert "add getResolve to webpack loader context (#7562)"
This reverts commit 5248117e657b2fb9833ac361953971cb08cd34fc.
