+++ crates/turbopack-node/Cargo.toml
@@ -19,7 +19,6 @@ workspace = true
 [dependencies]
 anyhow = { workspace = true }
 async-stream = "0.3.4"
-async-trait = { workspace = true }
 bytes = { workspace = true }
 const_format = "0.2.30"
 futures = { workspace = true }
@@ -44,7 +43,6 @@ turbopack-cli-utils = { workspace = true }
 turbopack-core = { workspace = true }
 turbopack-dev-server = { workspace = true }
 turbopack-ecmascript = { workspace = true }
-turbopack-resolve = { workspace = true }
 url = { workspace = true }
 urlencoding = { workspace = true }
