+++ crates/turbopack-node/js/src/ipc/evaluate.ts
@@ -1,136 +1,64 @@
 import { IPC, StructuredError } from "./index";
 import type { Ipc as GenericIpc } from "./index";
 
-type IpcIncomingMessage =
-  | {
-      type: "evaluate";
-      args: string[];
-    }
-  | {
-      type: "result";
-      id: number;
-      error: string | null;
-      data: any | null;
-    };
+type IpcIncomingMessage = {
+  type: "evaluate";
+  args: string[];
+};
 
 type IpcOutgoingMessage =
   | {
+      type: "value";
+      data: string;
+    }
+  | {
       type: "end";
       data: string | undefined;
-      duration: number;
     }
   | {
-      type: "info";
-      data: any;
+      type: "fileDependency";
+      path: string;
     }
   | {
-      type: "request";
-      id: number;
-      data: any;
+      type: "buildDependency";
+      path: string;
+    }
+  | {
+      type: "dirDependency";
+      path: string;
+      glob: string;
+    }
+  | {
+      type: "emittedError";
+      severity: "warning" | "error";
+      error: StructuredError;
     };
 
-export type Ipc<IM, RM> = {
-  sendInfo(message: IM): Promise<void>;
-  sendRequest(message: RM): Promise<unknown>;
-  sendError(error: Error): Promise<never>;
-};
-const ipc = IPC as GenericIpc<IpcIncomingMessage, IpcOutgoingMessage>;
-
-const queue: string[][] = [];
+export type Ipc = GenericIpc<IpcIncomingMessage, IpcOutgoingMessage>;
+const ipc = IPC as Ipc;
 
 export const run = async (
-  moduleFactory: () => Promise<{
-    init?: () => Promise<void>;
-    default: (ipc: Ipc<any, any>, ...deserializedArgs: any[]) => any;
-  }>
+  getValue: (ipc: Ipc, ...deserializedArgs: any[]) => any
 ) => {
-  let nextId = 1;
-  const requests = new Map();
-  const internalIpc = {
-    sendInfo: (message: any) =>
-      ipc.send({
-        type: "info",
-        data: message,
-      }),
-    sendRequest: (message: any) => {
-      const id = nextId++;
-      let resolve, reject;
-      const promise = new Promise((res, rej) => {
-        resolve = res;
-        reject = rej;
-      });
-      requests.set(id, { resolve, reject });
-      return ipc
-        .send({ type: "request", id, data: message })
-        .then(() => promise);
-    },
-    sendError: (error: Error) => {
-      return ipc.sendError(error);
-    },
-  };
-
-  // Initialize module and send ready message
-  let getValue: (ipc: Ipc<any, any>, ...deserializedArgs: any[]) => any;
-  try {
-    const module = await moduleFactory();
-    if (typeof module.init === "function") {
-      await module.init();
-    }
-    getValue = module.default;
-   await ipc.sendReady();
-  } catch (err) {
-    await ipc.sendReady();
-    await ipc.sendError(err as Error);
-  }
-
-
-  // Queue handling
-  let isRunning = false;
-  const run = async () => {
-    while (queue.length > 0) {
-      const args = queue.shift()!;
-      try {
-        const value = await getValue(internalIpc, ...args);
-        await ipc.send({
-          type: "end",
-          data:
-            value === undefined ? undefined : JSON.stringify(value, null, 2),
-          duration: 0,
-        });
-      } catch (e) {
-        await ipc.sendError(e as Error);
-      }
-    }
-    isRunning = false;
-  };
-
-  // Communication handling
   while (true) {
     const msg = await ipc.recv();
 
     switch (msg.type) {
       case "evaluate": {
-        queue.push(msg.args);
-        if (!isRunning) {
-          isRunning = true;
-          run();
-        }
-        break;
-      }
-      case "result": {
-        const request = requests.get(msg.id);
-        if (request) {
-          requests.delete(msg.id);
-          if (msg.error) {
-            request.reject(new Error(msg.error));
-          } else {
-            request.resolve(msg.data);
-          }
+        try {
+          const value = await getValue(ipc, ...msg.args);
+          await ipc.send({
+            type: "end",
+            data:
+              value === undefined ? undefined : JSON.stringify(value, null, 2),
+          });
+        } catch (e) {
+          await ipc.sendError(e as Error);
         }
         break;
       }
       default: {
-        console.error("unexpected message type", (msg as any).type);
+        console.error("unexpected message type", msg.type);
         process.exit(1);
       }
     }
