+++ crates/turbopack-node/js/src/transforms/postcss.ts
@@ -6,7 +6,6 @@ import postcss from "@vercel/turbopack/postcss";
 import importedConfig from "CONFIG";
 import { relative, isAbsolute, sep } from "path";
 import type { Ipc } from "../ipc/evaluate";
-import type { IpcInfoMessage, IpcRequestMessage } from "./webpack-loaders";
 
 const contextDir = process.cwd();
 
@@ -20,9 +19,11 @@ function toPath(file: string) {
   return sep !== "/" ? relPath.replaceAll(sep, "/") : relPath;
 }
 
-let processor: any;
-
-export const init = async (ipc: Ipc<IpcInfoMessage, IpcRequestMessage>) => {
+export default async function transform(
+  ipc: Ipc,
+  cssContent: string,
+  name: string
+) {
   let config = importedConfig;
   if (typeof config === "function") {
     config = await config({ env: "development" });
@@ -66,14 +67,7 @@ export const init = async (ipc: Ipc<IpcInfoMessage, IpcRequestMessage>) => {
     return plugin;
   });
 
-  processor = postcss(loadedPlugins);
-};
-
-export default async function transform(
-  ipc: Ipc<IpcInfoMessage, IpcRequestMessage>,
-  cssContent: string,
-  name: string
-) {
+  const processor = postcss(loadedPlugins);
   const { css, map, messages } = await processor.process(cssContent, {
     from: name,
     to: name,
@@ -98,26 +92,26 @@ export default async function transform(
         break;
       case "file-dependency":
       case "missing-dependency":
-        ipc.sendInfo({
+        ipc.send({
           type: "fileDependency",
           path: toPath(msg.file),
         });
         break;
       case "build-dependency":
-        ipc.sendInfo({
+        ipc.send({
           type: "buildDependency",
           path: toPath(msg.file),
         });
         break;
       case "dir-dependency":
-        ipc.sendInfo({
+        ipc.send({
           type: "dirDependency",
           path: toPath(msg.dir),
           glob: msg.glob,
         });
         break;
       case "context-dependency":
-        ipc.sendInfo({
+        ipc.send({
           type: "dirDependency",
           path: toPath(msg.file),
           glob: "**",
