+++ crates/turbopack/src/module_options/mod.rs
@@ -20,9 +20,7 @@ use turbopack_mdx::MdxTransformOptions;
 use turbopack_node::transforms::{postcss::PostCssTransform, webpack::WebpackLoaders};
 use turbopack_wasm::source::WebAssemblySourceType;
 
-use crate::{
-    evaluate_context::node_evaluate_asset_context, resolve_options_context::ResolveOptionsContext,
-};
+use crate::evaluate_context::node_evaluate_asset_context;
 
 #[turbo_tasks::function]
 async fn package_import_map_from_import_mapping(
@@ -61,7 +59,6 @@ impl ModuleOptions {
     pub async fn new(
         path: Vc<FileSystemPath>,
         module_options_context: Vc<ModuleOptionsContext>,
-        resolve_options_context: Vc<ResolveOptionsContext>,
     ) -> Result<Vc<ModuleOptions>> {
         let ModuleOptionsContext {
             enable_jsx,
@@ -88,11 +85,7 @@ impl ModuleOptions {
 
             for (condition, new_context) in rules.iter() {
                 if condition.matches(&path_value).await? {
-                    return Ok(ModuleOptions::new(
-                        path,
-                        *new_context,
-                        resolve_options_context,
-                    ));
+                    return Ok(ModuleOptions::new(path, *new_context));
                 }
             }
         }
@@ -555,7 +548,6 @@ impl ModuleOptions {
                                 execution_context,
                                 rule.loaders,
                                 rule.rename_as.clone(),
-                                resolve_options_context,
                             ),
                         )])),
                     ],
